refactor: removed unist_model
